Stardent ANSI lint
svn:revision:6377
cvs:account:mike
cvs:branch:trunk

+++ libutahrle/inv_cmap.c
@@ -754,7 +754,7 @@ maxfill( buffer, side )
 unsigned long *buffer;
 long side;
 {
-    register unsigned long maxv = ~0L;
+    register unsigned long maxv = (unsigned long)~0L;
     register long i;
     register unsigned long *bp;
